mageia: do NOT restart pbs_mom on upgrade, if jobs are running
The norm is that torque will have long-running jobs, which we definitely
not want to interrupt. Replacing the RPM could happen at any time, even
through some automated deployment tool. In that case, the usual "restart"
shall NOT interrupt our running jobs.

Instead, a "force-restart" is added to the init script, which will kill
the MOM at all cases.