scheduler.cc: raise username length to PBS_MAXUSER
Warning: breaks compatibility with previous scheduling usage data!

The allocated name[9] was too small for usernames > 8 bytes, which
would overflow the buffer or just corrupt the usage count. We'd
better be consistent with the rest of PBS and allow the maximum
username's length there.

But this means the old "/var/spool/torque/sched_priv/usage" file
will fail to be read() now. You have to reset it and let the
usage counts start from scratch again.